cli/tests/projects/functions/functions-python-v2/requirements.txt
azure-functions>=1.18.0,<2.0.0
